+++ pkg/service/core/api/gcp/gcp_service_account.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"time"
 
 	"github.com/navikt/nada-backend/pkg/errs"
 	"github.com/navikt/nada-backend/pkg/sa"
@@ -207,6 +208,8 @@ func (a *serviceAccountAPI) ensureServiceAccountExists(ctx context.Context, req 
 	if err != nil {
 		return nil, errs.E(errs.IO, service.CodeGCPServiceAccount, op, fmt.Errorf("creating service account '%s': %w", sa.ServiceAccountNameFromAccountID(req.ProjectID, req.AccountID), err))
 	}
+
+	time.Sleep(time.Minute)
 
 	return &service.ServiceAccountMeta{
 		Description: account.Description,
